+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.5.0-beta.x
+
+- Allow `BTreeMap` to be initialized with a `Record<string, any>` object (in addition to `Map`)
+- Allow for `HashMap<KeyType, ValueType>` definitions
+- Refresh the `NetworkState` types as returned by the `rpc.system.networkState()` call to current
+
 # 1.4.1 Feb 26, 2020
 
 - **Breaking change** (TypeScript only) The `*.d.ts` files now contain TypeScript 3.8 features (specifically `#private`), which is not usable in TS versions < 3.8